+++ src/handlebars-handlers/add-bonus-to-item-sheet.mjs
@@ -1,3 +1,4 @@
+import { MODULE_NAME } from '../consts.mjs';
 import { api } from '../util/api.mjs';
 import { handleJournalClick } from '../util/handle-journal-click.mjs';
 import { localize } from '../util/localize.mjs';
@@ -66,16 +67,44 @@ const addNodeToRollBonus = (itemSheetHtml, child, item, canEdit, inputType) => {
     if (!container) return;
 
     let bonusContainer = null;
-    if (inputType) {
+    if (inputType && item) {
         let bonusContainer = container.querySelector(`.${inputType}`);
         if (!bonusContainer) {
             bonusContainer = document.createElement('div');
             bonusContainer.classList.add(inputType);
             const header = document.createElement('h2');
-            header.classList.add('bonus-header');
             header.innerText = localize(`bonus-header-labels.${inputType}`);
-            container.appendChild(header);
+            header.classList.add('bonus-header');
+
+            if (inputType === 'target') {
+                const headerContainer = document.createElement('div');
+                headerContainer.classList.add('target-header');
+                headerContainer.appendChild(header);
+                bonusContainer.append(headerContainer);
+            }
+            else {
+                bonusContainer.appendChild(header);
+            }
+
             container.appendChild(bonusContainer);
+        }
+        // if target header already exists, then add target all/any toggle because this is the second target on it
+        else if (inputType === 'target') {
+            const hasToggle = !!bonusContainer.querySelector('.target-toggle');
+            const targetsOnItem = item[MODULE_NAME].targets.length;
+            if (!hasToggle && targetsOnItem >= 2) {
+                const headerContainer = bonusContainer.querySelector('.target-header');
+                if (headerContainer) {
+                    const toggle = createTemplate(
+                        templates.targetHeaderToggle,
+                        {
+                            readonly: !canEdit,
+                            current: item.getFlag(MODULE_NAME, 'target-toggle') || 'all',
+                        },
+                    );
+                    headerContainer.appendChild(toggle);
+                }
+            }
         }
     }