+++ src/target-and-bonus-join.mjs
@@ -115,11 +115,12 @@ export const handleBonusesFor = (thing, func, { skipGenericTarget = false } = {}
         .flatMap((targetType) => targetType.getSourcesFor(thing))
         // filter down to unique items in case one source item is affecting this target item through multiple "targets"
         .filter((sourceItem, i, self) => self.findIndex((nestedTarget) => sourceItem.id === nestedTarget.id) === i)
-        .filter((sourceItem) =>
-            sourceItem[MODULE_NAME].targets
-                .filter((sourceTarget) => !skipGenericTarget || !sourceTarget.isGenericTarget)
-                .every((sourceTarget) => sourceTarget.doesTargetInclude(sourceItem, thing))
-        )
+        .filter((sourceItem) => {
+            const targets = sourceItem[MODULE_NAME].targets
+                .filter((sourceTarget) => !skipGenericTarget || !sourceTarget.isGenericTarget);
+            const func = sourceItem.getFlag(MODULE_NAME, 'target-toggle') === 'all' ? 'every' : 'some';
+            return targets[func]((sourceTarget) => sourceTarget.doesTargetInclude(sourceItem, thing));
+        })
         .forEach((sourceItem) => sourceItem[MODULE_NAME].bonuses.forEach((bonusType) => func(bonusType, sourceItem)));
 }
 
@@ -137,10 +138,12 @@ export const handleBonusTypeFor = (thing, specificBonusType, func, { skipGeneric
         .flatMap((targetType) => targetType.getSourcesFor(thing))
         // filter down to unique items in case one source item is affecting this target item through multiple "targets"
         .filter((sourceItem, i, self) => self.findIndex((nestedTarget) => sourceItem.id === nestedTarget.id) === i)
-        .filter((sourceItem) => sourceItem[MODULE_NAME].targets
-            .filter((sourceTarget) => !skipGenericTarget || !sourceTarget.isGenericTarget)
-            .every((sourceTarget) => sourceTarget.doesTargetInclude(sourceItem, thing))
-        )
+        .filter((sourceItem) => {
+            const targets = sourceItem[MODULE_NAME].targets
+                .filter((sourceTarget) => !skipGenericTarget || !sourceTarget.isGenericTarget)
+            const func = sourceItem.getFlag(MODULE_NAME, 'target-toggle') === 'all' ? 'every' : 'some';
+            return targets[func]((sourceTarget) => sourceTarget.doesTargetInclude(sourceItem, thing));
+        })
         .forEach((sourceItem) => sourceItem[MODULE_NAME].bonuses.forEach((bonusType) => {
             if (bonusType === specificBonusType) {
                 func(specificBonusType, sourceItem);